fix scalar

+++ server/visualdl/lib.py
@@ -126,7 +126,7 @@ def get_invididual_image(storage, mode, tag, step_index, max_size=80):
             size = max(shape[0], shape[1])
             if size > max_size:
                 scale = max_size * 1. / size
-                scaled_shape = (shape[0]*scale, shape[1]*scale)
+                scaled_shape = (int(shape[0]*scale), int(shape[1]*scale))
                 im = im.resize(scaled_shape)
             im.save(tempfile)
         tempfile.seek(0, 0)
